project/requirements.txt
Flask>=3.0,<4.0
Pillow>=10.0,<12.0
numpy>=1.26,<3.0
requests>=2.31,<3.0
